Re-arrange router tests, ExcludeRegular-tag the big ones
