+++ src/test/scala/beam/sflight/SfLightRouterSpec.scala
@@ -5,8 +5,8 @@ import beam.agentsim.agents.vehicles.BeamVehicleType
 import beam.agentsim.agents.vehicles.VehicleProtocol.StreetVehicle
 import beam.agentsim.events.SpaceTime
 import beam.router.BeamRouter._
-import beam.router.Modes.BeamMode.{BIKE, CAR, RIDE_HAIL, WALK}
-import beam.router.model.{BeamLeg, BeamPath, BeamTrip}
+import beam.router.Modes.BeamMode.{BIKE, CAR, RIDE_HAIL, WALK, WALK_TRANSIT}
+import beam.router.model.{BeamLeg, BeamPath, BeamTrip, EmbodiedBeamTrip}
 import beam.router.{BeamRouter, Modes}
 import org.matsim.api.core.v01.{Coord, Id}
 import org.scalatest._
@@ -38,6 +38,35 @@ class SfLightRouterSpec extends AbstractSfLightSpec("SfLightRouterSpec") with In
       val walkTrip = response.itineraries.find(_.tripClassifier == WALK).getOrElse(fail)
       val routedStartTime = walkTrip.beamLegs.head.startTime
       assert(routedStartTime == time)
+    }
+
+    "respond with a transit route to a second reasonable RoutingRequest" in {
+      val origin = services.geo.wgs2Utm(new Coord(-122.396944, 37.79288)) // Embarcadero
+      val destination = services.geo.wgs2Utm(new Coord(-122.460555, 37.764294)) // Near UCSF medical center
+      val time = 25740
+      router ! RoutingRequest(
+        origin,
+        destination,
+        time,
+        withTransit = true,
+        Vector(
+          StreetVehicle(
+            Id.createVehicleId("body-667520-0"),
+            Id.create("BODY-TYPE-DEFAULT", classOf[BeamVehicleType]),
+            new SpaceTime(origin, time),
+            WALK,
+            asDriver = true
+          )
+        )
+      )
+      val response = expectMsgType[RoutingResponse]
+
+      assert(response.itineraries.exists(_.tripClassifier == WALK))
+      assert(response.itineraries.exists(_.tripClassifier == WALK_TRANSIT))
+      val transitOption = response.itineraries.find(_.tripClassifier == WALK_TRANSIT).get
+      assertMakesSense(transitOption.toBeamTrip)
+      assert(transitOption.costEstimate == 2.75)
+      assert(transitOption.legs.head.beamLeg.startTime == 25992)
     }
 
     "respond with fast travel time for a fast bike" in {
@@ -215,87 +244,58 @@ class SfLightRouterSpec extends AbstractSfLightSpec("SfLightRouterSpec") with In
 
     }
 
-    "respond with a walk route and usually a car route for each trip in sflight" in {
-      var numFailedCarRoutes = 0
-      scenario.getPopulation.getPersons
-        .values()
-        .forEach(person => {
-          val activities = planToVec(person.getSelectedPlan)
-          activities
-            .sliding(2)
-            .foreach(pair => {
-              val origin = pair(0).getCoord
-              val destination = pair(1).getCoord
-              val time = pair(0).getEndTime.toInt
-              router ! RoutingRequest(
-                origin,
-                destination,
-                time,
-                withTransit = false,
-                Vector(
-                  StreetVehicle(
-                    Id.createVehicleId("116378-2"),
-                    Id.create("Car", classOf[BeamVehicleType]),
-                    new SpaceTime(origin, 0),
-                    Modes.BeamMode.CAR,
-                    asDriver = true
-                  ),
-                  StreetVehicle(
-                    Id.createVehicleId("body-116378-2"),
-                    Id.create("Car", classOf[BeamVehicleType]),
-                    new SpaceTime(new Coord(origin.getX, origin.getY), time),
-                    Modes.BeamMode.WALK,
-                    asDriver = true
-                  )
-                )
-              )
-              val response = expectMsgType[RoutingResponse]
-              assert(response.itineraries.exists(_.tripClassifier == WALK))
+    "respond with a unlimited transfer route having cost 2.75 USD." in {
+      val origin = new Coord(549598.9574660371, 4176177.2431860007)
+      val destination = new Coord(544417.3891361314, 4177016.733758491)
+      val time = 64080
+      router ! RoutingRequest(
+        origin,
+        destination,
+        time,
+        withTransit = true,
+        Vector(
+          StreetVehicle(
+            Id.createVehicleId("body-667520-0"),
+            Id.create("BODY-TYPE-DEFAULT", classOf[BeamVehicleType]),
+            new SpaceTime(origin, time),
+            WALK,
+            asDriver = true
+          )
+        )
+      )
+      val response = expectMsgType[RoutingResponse]
 
-              val walkTrip = response.itineraries
-                .find(_.tripClassifier == WALK)
-                .get
-                .toBeamTrip
-              inside(walkTrip) {
-                case BeamTrip(legs) =>
-                  legs.map(_.mode) should contain theSameElementsInOrderAs List(WALK)
-                  inside(legs.loneElement) {
-                    case BeamLeg(_, mode, _, BeamPath(_, _, _, _, _, _)) =>
-                      mode should be(WALK)
-                  }
-              }
-
-              if (response.itineraries.exists(_.tripClassifier == CAR)) {
-                val carTrip = response.itineraries
-                  .find(_.tripClassifier == CAR)
-                  .get
-                  .toBeamTrip
-                assertMakesSense(carTrip)
-                inside(carTrip) {
-                  case BeamTrip(legs) =>
-                    legs should have size 3
-                    inside(legs(0)) {
-                      case BeamLeg(_, mode, _, BeamPath(_, _, _, _, _, _)) =>
-                        mode should be(WALK)
-                    }
-                    inside(legs(1)) {
-                      case BeamLeg(_, mode, _, BeamPath(links, _, _, _, _, _)) =>
-                        mode should be(CAR)
-                        links should not be 'empty
-                    }
-                    inside(legs(2)) {
-                      case BeamLeg(_, mode, _, BeamPath(_, _, _, _, _, _)) =>
-                        mode should be(WALK)
-                    }
-                }
-              } else {
-                numFailedCarRoutes = numFailedCarRoutes + 1
-              }
-            })
-        })
-      // Sometimes car routes fail, but should be very rare
-      assert(numFailedCarRoutes < 7)
+      assert(response.itineraries.exists(_.costEstimate == 2.75))
+      assert(response.itineraries.exists(_.tripClassifier == WALK))
+      assert(response.itineraries.exists(_.tripClassifier == WALK_TRANSIT))
     }
+
+    "respond with a BART route without transfer having cost 1.95 USD." in {
+      val origin = services.geo.wgs2Utm(new Coord(-122.41969, 37.76506)) // 16th St. Mission
+      val destination = services.geo.wgs2Utm(new Coord(-122.40686, 37.784992)) // Powell St.
+      val time = 51840
+      router ! RoutingRequest(
+        origin,
+        destination,
+        time,
+        withTransit = true,
+        Vector(
+          StreetVehicle(
+            Id.createVehicleId("body-667520-0"),
+            Id.create("BODY-TYPE-DEFAULT", classOf[BeamVehicleType]),
+            new SpaceTime(origin, time),
+            WALK,
+            asDriver = true
+          )
+        )
+      )
+      val response = expectMsgType[RoutingResponse]
+
+      assert(response.itineraries.exists(_.costEstimate == 1.95))
+      assert(response.itineraries.exists(_.tripClassifier == WALK))
+      assert(response.itineraries.exists(_.tripClassifier == WALK_TRANSIT))
+    }
+
   }
 
   def assertMakesSense(trip: BeamTrip): Unit = {
